--- src/recoil/imgModal.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const imgModalState = atom<boolean>({
-  key: 'imgModal',
-  default: false,
-});
