+++ src/recoil/index.ts
@@ -4,7 +4,7 @@ import { reviewmodalState, tagitem, starRate, textRate } from './reviewwritemoda
 import { ReviewState, MyReviewState } from './Reviewlist';
 import { InputState } from './Owner/owner';
 import { searchToggleState, sellingProductSearchToggle } from './searchToggleState';
-import { imgModalState } from './imgModal';
+import { imgModalState, profileImgState, nickNameState } from './editState';
 
 export {
   exampleState,
@@ -19,4 +19,6 @@ export {
   MyReviewState,
   sellingProductSearchToggle,
   imgModalState,
+  profileImgState,
+  nickNameState,
 };
